requirements.txt
pyyaml
AWSIoTPythonSDK
